refactor: extract pure data types from control to entity packages (Phase 6)
Move pure data types (case classes, enums) from control packages into their
domain entity packages to break false coupling. Types moved:
- ControlPlaneModels (24 types) → orchestration-domain entity
- WorkspaceAnalysisState/Status → analysis-domain entity
- ProviderAvailability, AuthStatus, etc. → config-domain entity
- KnowledgeEdge, KnowledgeDecisionMatch, etc. → knowledge-domain entity
- GatewayQueueCommand, ChannelMetrics, etc. → gateway-domain entity
- AssignRunRequest, RequirementCheck → workspace-domain entity
