+++ src/main/scala/conversation/boundary/ChatController.scala
@@ -13,7 +13,8 @@ import activity.control.ActivityHub
 import activity.entity.{ ActivityEvent, ActivityEventType }
 import conversation.entity.api.*
 import db.{ ChatRepository, TaskRepository }
-import gateway.control.{ ChannelRegistry, GatewayService, GatewayServiceError, MessageChannelError }
+import gateway.control.{ ChannelRegistry, GatewayService, MessageChannelError }
+import gateway.control.GatewayServiceError
 import gateway.entity.{ GatewayMessageRole as GatewayMessageRole, MessageDirection as GatewayMessageDirection, * }
 import issues.entity.{ IssueReport, IssueWorkReport }
 import llm4zio.core.{ ConversationThread, LlmError, LlmService, Streaming, ToolConversationManager }
