+++ src/main/scala/gateway/boundary/ChannelController.scala
@@ -10,7 +10,7 @@ import zio.json.*
 
 import _root_.config.entity.{ ConfigRepository, GatewayConfig }
 import gateway.control.*
-import gateway.entity.SessionScopeStrategy
+import gateway.entity.{ ChannelMetrics, SessionScopeStrategy }
 import shared.errors.PersistenceError
 import shared.web.{ ChannelCardData, ChannelView, HtmlViews }
