+++ src/main/scala/gateway/control/GatewayService.scala
@@ -3,10 +3,10 @@ package gateway.control
 import java.util.UUID
 
 import zio.*
-import zio.json.*
 
 import _root_.config.entity.{ AgentInfo, ConfigRepository }
 import db.ChatRepository
+import gateway.entity.*
 import gateway.entity.NormalizedMessage
 import llm4zio.core.{ LlmService, Streaming }
 import memory.entity.{ MemoryEntry, MemoryId, MemoryKind, MemoryRepository }
@@ -16,29 +16,6 @@ import prompts.PromptLoader
 enum GatewayServiceError:
   case Router(error: MessageRouterError)
   case QueueClosed
-
-enum GatewayQueueCommand:
-  case Inbound(message: NormalizedMessage)
-  case Outbound(message: NormalizedMessage)
-
-case class ChannelMetrics(
-  inboundEnqueued: Long = 0L,
-  outboundEnqueued: Long = 0L,
-  inboundProcessed: Long = 0L,
-  outboundProcessed: Long = 0L,
-  failed: Long = 0L,
-  lastActivityTs: Option[Long] = None,
-) derives JsonCodec
-
-case class GatewayMetricsSnapshot(
-  enqueued: Long = 0L,
-  processed: Long = 0L,
-  failed: Long = 0L,
-  chunkedMessages: Long = 0L,
-  emittedChunks: Long = 0L,
-  steeringForwarded: Long = 0L,
-  perChannel: Map[String, ChannelMetrics] = Map.empty,
-)
 
 trait GatewayService:
   def enqueueInbound(message: NormalizedMessage): UIO[Unit]
