+++ src/main/scala/gateway/control/MessageRouter.scala
@@ -10,6 +10,7 @@ import conversation.entity.api.StoredSessionContext
 import db.ChatRepository
 import gateway.entity.*
 import orchestration.control.*
+import orchestration.entity.*
 import shared.errors.PersistenceError
 
 enum MessageRouterError:
