+++ src/main/scala/orchestration/control/AutoDispatcher.scala
@@ -12,7 +12,8 @@ import governance.entity.{ GovernanceLifecycleAction, GovernanceLifecycleStage, 
 import issues.entity.{ AgentIssue, IssueEvent, IssueRepository, IssueState }
 import shared.errors.PersistenceError
 import shared.ids.Ids.{ AgentId, EventId, TaskRunId }
-import workspace.control.{ AssignRunRequest, WorkspaceRunService }
+import workspace.control.WorkspaceRunService
+import workspace.entity.AssignRunRequest
 import workspace.entity.WorkspaceRepository
 
 trait AutoDispatcher:
