+++ src/main/scala/orchestration/control/TaskExecutor.scala
@@ -5,6 +5,7 @@ import java.util.UUID
 import zio.*
 
 import _root_.config.entity.{ WorkflowContext, WorkflowDefinition }
+import orchestration.entity.*
 import db.TaskRepository
 import shared.errors.PersistenceError
 import taskrun.entity.RunStatus
