+++ src/test/scala/analysis/control/WorkspaceAnalysisSchedulerSpec.scala
@@ -7,7 +7,7 @@ import zio.test.{ Live, * }
 
 import _root_.config.entity.SettingRow
 import activity.entity.{ ActivityEvent, ActivityEventType }
-import analysis.entity.{ AnalysisDoc, AnalysisEvent, AnalysisRepository, AnalysisType }
+import analysis.entity.{ AnalysisDoc, AnalysisEvent, AnalysisRepository, AnalysisType, WorkspaceAnalysisState, WorkspaceAnalysisStatus }
 import board.entity.*
 import db.*
 import project.control.ProjectStorageService
