+++ src/test/scala/gateway/MessageRouterSpec.scala
@@ -11,7 +11,8 @@ import conversation.entity.api.{ ChatConversation, ConversationEntry, SessionCon
 import db.*
 import gateway.control.*
 import gateway.entity.*
-import orchestration.control.{ OrchestratorControlPlane, WorkflowFailed }
+import orchestration.control.OrchestratorControlPlane
+import orchestration.entity.WorkflowFailed
 import shared.errors.PersistenceError
 
 object MessageRouterSpec extends ZIOSpecDefault:
