+++ src/test/scala/models/ModelsCoverageSpec.scala
@@ -8,7 +8,7 @@ import zio.test.*
 import _root_.config.entity.{ AgentHealth, AgentHealthStatus, AgentMetrics }
 import conversation.entity.api.{ ChatConversationCreateRequest, ConversationMessageRequest, MessageType }
 import issues.entity.api.{ AgentAssignmentView, AgentIssueCreateRequest, AgentIssueView, IssuePriority }
-import orchestration.control.*
+import orchestration.entity.*
 import shared.errors.{ ControlPlaneError, WorkspaceError }
 import taskrun.entity.{ ProgressModels, StepMetrics, StepProgressEvent }
