+++ src/test/scala/project/boundary/ProjectsControllerSpec.scala
@@ -7,7 +7,8 @@ import zio.http.*
 import zio.test.*
 
 import _root_.config.entity.*
-import analysis.control.{ WorkspaceAnalysisScheduler, WorkspaceAnalysisState, WorkspaceAnalysisStatus }
+import analysis.control.WorkspaceAnalysisScheduler
+import analysis.entity.{ WorkspaceAnalysisState, WorkspaceAnalysisStatus }
 import analysis.entity.AnalysisType
 import issues.entity.{ AgentIssue, IssueState }
 import orchestration.control.AgentRegistry
