+++ src/test/scala/shared/web/AgentMonitorViewSpec.scala
@@ -5,7 +5,7 @@ import java.time.Instant
 import zio.test.*
 
 import llm4zio.observability.MetricsSnapshot
-import orchestration.control.{ AgentExecutionInfo, AgentExecutionState, AgentMonitorSnapshot }
+import orchestration.entity.{ AgentExecutionInfo, AgentExecutionState, AgentMonitorSnapshot }
 import shared.web.AgentMonitorView.{ AgentGlobalStats, AgentRunView }
 
 object AgentMonitorViewSpec extends ZIOSpecDefault:
