+++ src/test/scala/shared/web/KnowledgeViewSpec.scala
@@ -5,7 +5,7 @@ import java.time.Instant
 import zio.test.*
 
 import analysis.entity.{ AnalysisDoc, AnalysisType }
-import knowledge.control.{ ArchitecturalContext, KnowledgeDecisionMatch, KnowledgeEdge }
+import knowledge.entity.{ ArchitecturalContext, KnowledgeDecisionMatch, KnowledgeEdge }
 import knowledge.entity.{ DecisionLog, DecisionLogVersion, DecisionMaker, DecisionMakerKind }
 import memory.entity.{ MemoryEntry, MemoryId, MemoryKind, SessionId, UserId }
 import shared.ids.Ids.{ AgentId, DecisionLogId, IssueId }
