+++ src/test/scala/web/controllers/AgentMonitorControllerSpec.scala
@@ -10,6 +10,7 @@ import _root_.config.entity.WorkflowDefinition
 import app.boundary.AgentMonitorControllerLive
 import llm4zio.observability.LlmMetrics
 import orchestration.control.*
+import orchestration.entity.*
 import shared.errors.ControlPlaneError
 import taskrun.entity.TaskStep
