+++ src/test/scala/web/controllers/IssueControllerSpec.scala
@@ -28,7 +28,7 @@ import shared.errors.PersistenceError
 import shared.ids.Ids.{ AgentId, AnalysisDocId, BoardIssueId, IssueId }
 import shared.testfixtures.*
 import taskrun.entity.{ TaskArtifactRow, TaskReportRow, TaskRunRow }
-import workspace.control.{ AssignRunRequest, WorkspaceRunService }
+import workspace.control.WorkspaceRunService
 import workspace.entity.*
 
 object IssueControllerSpec extends ZIOSpecDefault:
